+++ packages/core-js-compat/src/data.mjs
@@ -1464,8 +1464,12 @@ export const data = {
     bun: '1.1.1',
     // v8 ~ Chrome 122 does not properly work with set-like objects
     // https://bugs.chromium.org/p/v8/issues/detail?id=14559
-    chrome: '123',
+    // v8 < Chrome 128 does not properly convert set-like objects Infinty size
+    // https://issues.chromium.org/issues/351332634
+    chrome: '128', // '122',
     firefox: '127',
+    // https://github.com/nodejs/node/pull/54883
+    node: '22.10',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
     safari: '18.0', // '17.0',
@@ -1474,8 +1478,12 @@ export const data = {
     bun: '1.1.1',
     // v8 ~ Chrome 122 does not properly work with set-like objects
     // https://bugs.chromium.org/p/v8/issues/detail?id=14559
-    chrome: '123',
+    // v8 < Chrome 128 does not properly convert set-like objects Infinty size
+    // https://issues.chromium.org/issues/351332634
+    chrome: '128', // '122',
     firefox: '127',
+    // https://github.com/nodejs/node/pull/54883
+    node: '22.10',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
     safari: '18.0', // '17.0',
@@ -1484,8 +1492,12 @@ export const data = {
     bun: '1.1.1',
     // v8 ~ Chrome 122 does not properly work with set-like objects
     // https://bugs.chromium.org/p/v8/issues/detail?id=14559
-    chrome: '123',
+    // v8 < Chrome 128 does not properly convert set-like objects Infinty size
+    // https://issues.chromium.org/issues/351332634
+    chrome: '128', // '122',
     firefox: '127',
+    // https://github.com/nodejs/node/pull/54883
+    node: '22.10',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
     safari: '18.0', // '17.0',
@@ -1494,8 +1506,12 @@ export const data = {
     bun: '1.1.1',
     // v8 ~ Chrome 122 does not properly work with set-like objects
     // https://bugs.chromium.org/p/v8/issues/detail?id=14559
-    chrome: '123',
+    // v8 < Chrome 128 does not properly convert set-like objects Infinty size
+    // https://issues.chromium.org/issues/351332634
+    chrome: '128', // '122',
     firefox: '127',
+    // https://github.com/nodejs/node/pull/54883
+    node: '22.10',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
     safari: '18.0', // '17.0',
@@ -1504,8 +1520,12 @@ export const data = {
     bun: '1.1.1',
     // v8 ~ Chrome 122 does not properly work with set-like objects
     // https://bugs.chromium.org/p/v8/issues/detail?id=14559
-    chrome: '123',
+    // v8 < Chrome 128 does not properly convert set-like objects Infinty size
+    // https://issues.chromium.org/issues/351332634
+    chrome: '128', // '122',
     firefox: '127',
+    // https://github.com/nodejs/node/pull/54883
+    node: '22.10',
     // safari 17 does not properly work with set-like objects
     // https://bugs.webkit.org/show_bug.cgi?id=267494
     safari: '18.0', // '17.0',